pyproject.toml
[project]
name = "tessie_api"
version = "0.0.1"
authors = [{ name = "Andrew Gierens", email = "apgierens@gmail.com" }]
description = "A small python package that wraps up the Tessie API "
readme = "README.md"
requires-python = ">=3.5"
exclude = ["*.env", "*.env.example", ".devcontainer/", ".vscode/", "tests/"]
dependencies = [
    "PyJWT==2.8.0",
    "python-dotenv>=1.0.1",
    "aiohttp>=3.9.5",
    "pytz>=2025.1",
    "pytest-asyncio>=0.23.8",
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: OS Independent",
]

[project.urls]
"Homepage" = "https://github.com/andrewgierens/tessie_python_api"
"Bug Tracker" = "https://github.com/andrewgierens/tessie_python_api/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
